AEA-3830 Validate resourceType of Task resources.
